mybatis cache

+++ mybatis/src/main/java/com/albenyuan/mybatis/mapper/MenuMapper.java
@@ -8,13 +8,13 @@ import java.util.List;
 
 @Repository
 public interface MenuMapper {
-    int deleteByPrimaryKey(Integer id);
+    int deleteByPrimaryKey(Long id);
 
     int insert(Menu record);
 
     int insertSelective(Menu record);
 
-    Menu selectByPrimaryKey(Integer id);
+    Menu selectByPrimaryKey(Long id);
 
     int updateByPrimaryKeySelective(Menu record);
